# docs/api/admin-openapi.yaml
openapi: 3.0.3
info:
  title: IoLink Admin API — TARGET
  version: 1.0.0-planned
  description: TARGET / 待实现、待验收：本文件保留 M0–M5 管理 v1 契约，并包含 M6a products/models/assignment 增量；并非现有实现已通过验收的声明。原有 v1 操作路径保持不变；新增目标操作见各 summary。所有 JSON 字段 snake_case；集合始终返回数组，无结果为 []，不得返回 null。时间统一返回 UTC RFC3339
    字符串。所有参数越界/非法值返回 400，不静默回退。所有受保护操作必须逐请求校验身份及资源归属。管理员使用aid claim和管理签名密钥，小程序使用uid claim和用户签名密钥；均由IOLINK_SECRET_KEY以不同标签派生（逻辑密钥，不是新增环境变量），不得互验。管理权限覆盖本实例全部业务资源，管理员不会自动成为
    farm.owner_id。管理集合均支持 limit/offset，默认按 id 升序稳定排序；报警按 created_at DESC,id DESC。新增资源 POST 非幂等，网络结果未知时须查询核实，不能盲目重试；PUT 为完整字段替换且重复执行结果相同；确认操作及 DELETE 幂等，重复成功不产生新副作用。
servers:
- url: http://localhost:8080/admin/v1
...
security:
- BearerAuth: []
paths:
  /login:
    post:
      tags:
      - auth
      summary: 管理员登录
      description: 凭据错误统一 401。POST 非幂等，每次成功可签发新 JWT。
      responses:
        '200':
          description: 成功
          content:
            application/json:
              schema:
                type: object
                required:
                - token
                - expires_in
                properties:
                  token:
                    type: string
                  expires_in:
                    type: integer
                    minimum: 1
                    description: 秒
                additionalProperties: false
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '500':
          $ref: '#/components/responses/InternalError'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
              - username
              - password
              properties:
                username:
                  type: string
                  minLength: 1
                password:
                  type: string
                  minLength: 1
                  format: password
              additionalProperties: false
      security: []
  /tenants:
    get:
      tags: [tenants]
      summary: 租户列表或当前租户
      responses:
        '200':
          description: 成功
          content:
            application/json:
              schema:
                type: array
                items: { $ref: '#/components/schemas/Tenant' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '500': { $ref: '#/components/responses/InternalError' }
  /tenants/{id}/status:
    put:
      tags: [tenants]
      summary: 启用或停用租户（平台管理员）
      parameters:
      - { name: id, in: path, required: true, schema: { type: integer, format: int64, minimum: 1 } }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [active]
              properties: { active: { type: boolean } }
              additionalProperties: false
      responses:
        '204': { description: 成功 }
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
  /tenants/{id}/members:
    get:
      tags: [tenants]
      summary: 租户成员列表
      parameters:
      - { name: id, in: path, required: true, schema: { type: integer, format: int64, minimum: 1 } }
      responses:
        '200':
          description: 成功
          content:
            application/json:
              schema:
                type: array
                items: { $ref: '#/components/schemas/TenantMembership' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '404': { $ref: '#/components/responses/NotFound' }
  /tenants/{id}/members/{user_id}:
    put:
      tags: [tenants]
      summary: 更新租户成员角色或状态
      parameters:
      - { name: id, in: path, required: true, schema: { type: integer, format: int64, minimum: 1 } }
      - { name: user_id, in: path, required: true, schema: { type: integer, format: int64, minimum: 1 } }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [role, active]
              properties:
                role: { type: string, enum: [owner, admin, member, viewer, support] }
                active: { type: boolean }
                expires_at: { type: string, format: date-time, nullable: true }
                expires_at: { type: string, format: date-time, nullable: true }
              additionalProperties: false
      responses:
        '204': { description: 成功 }
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
  /password:
    post:
      tags:
      - auth
      summary: 修改自己的密码
      description: 校验旧密码，错误统一 401。成功后增加 token_version，全部既有管理员 token 即时失效，需要重新登录。非幂等：旧 token 或旧密码重试返回 401；不得在响应或日志暴露口令。
      responses:
        '204':
          description: 成功；响应无正文
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '500':
          $ref: '#/components/responses/InternalError'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
              - old_password
              - new_password
              properties:
                old_password:
                  type: string
                  minLength: 1
                  format: password
                new_password:
                  type: string
                  minLength: 8
                  format: password
              additionalProperties: false
  /users:
    get:
      tags:
      - users
      summary: '[新增目标] 搜索可分配养殖场的微信用户'
      description: 仅管理端可调用，只输出 id/nickname；不允许按 openid 搜索或返回敏感字段。
      responses:
        '200':
          description: 成功
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/User'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '500':
          $ref: '#/components/responses/InternalError'
      parameters:
      - name: query
        in: query
        required: true
        schema:
          type: string
          minLength: 1
        description: 按纯数字 id 精确匹配或 nickname 子串搜索；仅 USER 且已注册。
      - &id001
        name: limit
        in: query
        required: false
        schema:
          type: integer
          minimum: 1
          maximum: 200
          default: 50
      - &id002
        name: offset
        in: query
        required: false
        schema:
          type: integer
          minimum: 0
          default: 0
  /farms:
    get:
      tags:
      - farms
      summary: 养殖场列表
      description: 按目标契约实现并以集成测试验收。
      responses:
        '200':
          description: 成功
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Farm'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '500':
          $ref: '#/components/responses/InternalError'
      parameters:
      - *id001
      - *id002
    post:
      tags:
      - farms
      summary: 新建养殖场
      description: owner_id 缺省或 null 创建未分配养殖场；非空目标用户不存在或不是已注册 USER 返回 404。POST 非幂等。
      responses:
        '200':
          description: 成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Farm'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalError'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/FarmCreateReq'
  /farms/{id}:
    put:
      tags:
      - farms
      summary: 修改养殖场
      description: 幂等替换 name/location；省略 location 置空；不修改 owner_id。
      responses:
        '204':
          description: 成功；响应无正文
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalError'
      parameters:
      - &id003
        name: id
        in: path
        required: true
        schema:
          type: integer
          format: int64
          minimum: 1
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/FarmReq'
    delete:
      tags:
      - farms
      summary: 删除养殖场
      description: 无记录或已删除返回 204；仍有池塘或历史引用返回 409，不级联删除历史。
      responses:
        '204':
          description: 成功；响应无正文
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '409':
          $ref: '#/components/responses/Conflict'
        '500':
          $ref: '#/components/responses/InternalError'
      parameters:
      - *id003
  /farms/{id}/owner:
    put:
      tags:
      - farms
      summary: '[新增目标] 分配、转交或解除养殖场归属'
      description: 幂等。非空必须为已注册 USER，否则 404。null 解除分配：小程序无人可见、订阅通知暂停。事务提交后旧 owner 立即失去该场池塘及历史访问权；新 owner 获得该场现有及历史数据。管理员始终可管理。通知发送前重新核验当前 owner。
      responses:
        '204':
          description: 成功；响应无正文
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalError'
      parameters:
      - *id003
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
              - owner_id
              properties:
                owner_id:
                  type: integer
                  format: int64
                  minimum: 1
                  nullable: true
              additionalProperties: false
  /farms/{id}/members:
    get:
      tags: [farms]
      summary: 养殖场成员列表
      parameters:
      - *id003
      responses:
        '200':
          description: 成功
          content:
            application/json:
              schema:
                type: array
                items: { $ref: '#/components/schemas/FarmMembership' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '404': { $ref: '#/components/responses/NotFound' }
  /farms/{id}/members/{user_id}:
    put:
      tags: [farms]
      summary: 更新养殖场成员
      parameters:
      - *id003
      - { name: user_id, in: path, required: true, schema: { type: integer, format: int64, minimum: 1 } }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [role, active]
              properties:
                role: { type: string, enum: [owner, admin, member, viewer, support] }
                active: { type: boolean }
              additionalProperties: false
      responses:
        '204': { description: 成功 }
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '404': { $ref: '#/components/responses/NotFound' }
  /ponds:
    get:
      tags:
      - ponds
      summary: 池塘列表及最新状态
      description: 按目标契约实现并以集成测试验收。
      responses:
        '200':
          description: 成功
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Pond'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '500':
          $ref: '#/components/responses/InternalError'
      parameters:
      - name: farm_id
        in: query
        required: false
        schema:
          type: integer
          format: int64
          minimum: 1
      - *id001
      - *id002
    post:
      tags:
      - ponds
      summary: 新建池塘
      description: farm_id 不存在返回 404；成功初始 status=normal、device_count=0、latest=null。POST 非幂等。
      responses:
        '200':
          description: 成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Pond'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalError'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PondReq'
  /ponds/{id}:
    put:
      tags:
      - ponds
      summary: 修改池塘
      description: 幂等；farm_id 必须保持不变。省略 area_mu 置 0。
      responses:
        '204':
          description: 成功；响应无正文
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalError'
      parameters:
      - *id003
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PondUpdateReq'
    delete:
      tags:
      - ponds
      summary: 删除池塘
      description: 无记录或已删除返回204；有任何设备记录（含停用）、规则、历史遥测或报警引用返回409。M0–M5不提供绕过该限制的归档/硬删入口，不级联删除历史。
      responses:
        '204':
          description: 成功；响应无正文
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '409':
          $ref: '#/components/responses/Conflict'
        '500':
          $ref: '#/components/responses/InternalError'
      parameters:
      - *id003
  /devices:
    get:
      tags:
      - devices
      summary: 设备列表
      description: 默认返回未停用设备，include_disabled=true 包含停用设备；不返回明文 secret 或摘要。
      responses:
        '200':
          description: 成功
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Device'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '500':
          $ref: '#/components/responses/InternalError'
      parameters:
      - name: pond_id
        in: query
        required: false
        schema:
          type: integer
          format: int64
          minimum: 1
      - name: include_disabled
        in: query
        required: false
        schema:
          type: boolean
          default: false
      - *id001
      - *id002
    post:
      tags:
      - devices
      summary: 注册设备并返回一次性 secret
      description: 强制绑定已存在 pond_id；无裸设备/裸解绑状态。device_no 服务端生成且永久唯一。secret 只在本次响应显示，服务端仅存摘要。POST 非幂等，响应丢失不得自动重试；管理员核实后停用旧记录并重新注册。
      responses:
        '200':
          description: 成功
          content:
            application/json:
              schema:
                type: object
                required:
                - device_no
                - secret
                - pond_id
                - name
                - model
                - status
                - report_interval
                properties:
                  device_no:
                    type: string
                  secret:
                    type: string
                    minLength: 64
                    maxLength: 64
                    pattern: ^[0-9a-f]{64}$
                    description: 32字节随机值的64字符hex编码，仅本次返回。
                  pond_id: &id004
                    type: integer
                    format: int64
                    minimum: 1
                  name:
                    type: string
                  model:
                    type: string
                  status:
                    type: string
                    enum:
                    - offline
                  report_interval:
                    type: integer
                    enum:
                    - 60
                    - 300
                additionalProperties: false
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalError'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
              - pond_id
              properties:
                pond_id: *id004
                name:
                  type: string
                  default: ''
                model:
                  type: string
                  default: ''
                report_interval:
                  type: integer
                  enum:
                  - 60
                  - 300
                  description: 缺省继承服务器当时IOLINK_REPORT_INTERVAL默认值（60秒），写入设备记录；固件需同周期。
              additionalProperties: false
  /devices/{device_no}:
    get:
      tags:
      - devices
      summary: '[新增目标] 设备详情及最新数据'
      description: 管理端可读取停用设备详情及已保留影子；latest 只来自设备当前池塘的样本，没有样本为 null。
      responses:
        '200':
          description: 成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DeviceDetail'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalError'
      parameters:
      - &id005
        name: device_no
        in: path
        required: true
        schema:
          type: string
          minLength: 1
    delete:
      tags:
      - devices
      summary: 停用设备并保留历史
      description: 幂等；软删除保留历史遥测、报警和池塘绑定；撤销 MQTT 连接及重连权限，立即使其退出在线计数，不再接受新消息。重复停用或不存在均 204。device_no 不得复用；管理历史报警仍可查看和确认。
      responses:
        '204':
          description: 成功；响应无正文
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '500':
          $ref: '#/components/responses/InternalError'
      parameters:
      - *id005
...
  /alarms:
    get:
      tags:
      - alarms
      summary: 报警列表（管理视角）
      description: 按目标契约实现并以集成测试验收。
      responses:
        '200':
          description: 成功
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Alarm'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '500':
          $ref: '#/components/responses/InternalError'
      parameters:
      - name: level
        in: query
        required: false
        schema:
          type: string
          enum:
          - critical
          - warning
      - name: only_unconfirmed
        in: query
        required: false
        schema:
          type: boolean
          default: false
      - *id001
      - *id002
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  responses:
    BadRequest:
      description: 参数、JSON 或业务校验失败；error=invalid_request
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    Unauthorized:
      description: 凭据缺失/过期/错误、错误 token 类型或权限主体失效；error=unauthorized
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
...
    InternalError:
      description: 内部故障；error=internal_error
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
...
  schemas:
...
    Tenant:
      type: object
      required: [id, name, active, permission_version]
      properties:
        id: { type: integer, format: int64, minimum: 1 }
        name: { type: string, minLength: 1 }
        active: { type: boolean }
        permission_version: { type: integer, format: int64, minimum: 0 }
      additionalProperties: false
...